Make modules of all cmd/* tools
This change removes the movefunds and sweepaccount tools from the main module by
declaring both as their own modules.  Dividing the modules this way allows the
main module (to build the dcrwallet executable) to remove requires on
dependencies currently used by these tools.

The sweepaccount tool had to be modified to avoid importing any packages from
the main module.  The main module is not intended to be used as a requirement
and is only used for reproducible builds.